@@ -1,27 +1,27 @@
 {
   "name": "xcube-viewer",
-  "version": "0.4.0-dev.0",
+  "version": "0.3.2",
   "private": true,
   "homepage": ".",
   "scripts": {
-    "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build",
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject"
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
   },
   "dependencies": {
-    "@date-io/date-fns": "^1.3.9",
-    "@material-ui/core": "^4.4.0",
-    "@material-ui/lab": "^4.0.0-alpha.25",
-    "@material-ui/icons": "^4.2.1",
-    "@material-ui/pickers": "^3.2.4",
+    "@date-io/date-fns": "^1.3.13",
+    "@material-ui/core": "^4.8.0",
+    "@material-ui/icons": "^4.5.1",
+    "@material-ui/lab": "^4.0.0-alpha.36",
+    "@material-ui/pickers": "^3.2.8",
     "classnames": "^2.2.6",
     "date-fns": "^2.0.1",
     "ol": "^6.1.1",
-    "react": "^16.9.0",
-    "react-dom": "^16.9.0",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
     "react-redux": "^6.0.0",
-    "react-scripts-ts": "3.1.0",
+    "react-scripts": "3.3.0",
     "recharts": "^1.7.1",
     "redux": "^4.0.4",
     "redux-logger": "^3.0.6",
@@ -29,25 +29,40 @@
     "reselect": "^4.0.0"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/user-event": "^7.1.2",
     "@types/classnames": "^2.2.7",
-    "@types/enzyme": "^3.1.17",
-    "@types/enzyme-adapter-react-16": "^1.0.5",
     "@types/geojson": "^7946.0.5",
     "@types/jest": "^24.0.0",
+    "@types/node": "^12.0.0",
     "@types/ol": "^5.3.6",
-    "@types/node": "^10.12.21",
     "@types/react": "^16.9.2",
-    "@types/react-dom": "^16.0.11",
+    "@types/react-dom": "^16.9.0",
     "@types/react-redux": "^7.1.2",
     "@types/recharts": "=1.1.10",
     "@types/redux-logger": "^3.0.7",
     "@types/redux-thunk": "^2.1.0",
-    "enzyme": "^3.8.0",
-    "enzyme-adapter-react-16": "^1.9.1",
-    "typescript": "^3.6.2"
+    "typescript": "~3.7.2"
   },
   "resolutions": {
-    "react": "^16.9.0",
-    "@types/react": "^16.9.2"
+    "react": "^16.12.0",
+    "@types/react": "^16.9.2",
+    "set-value": "^2.0.1"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
