@@ -1,5 +1,5 @@
-import * as React from 'react';
-import * as enzyme from 'enzyme';
+import React from 'react';
+import { render } from '@testing-library/react';
 import ErrorBoundary from './ErrorBoundary';
 
 function MyWidget(props: { name: string | null }) {
@@ -11,19 +11,22 @@ function MyWidget(props: { name: string | null }) {
 
 describe('ErrorBoundary', () => {
     it('renders the child if child does not throw', () => {
-        const c = enzyme.mount(<ErrorBoundary><MyWidget name="Bibo"/></ErrorBoundary>);
-        expect(c.find('.myWidget').text()).toEqual('Hi Bibo!');
+        const {getByText} = render(<ErrorBoundary><MyWidget name="Bibo"/></ErrorBoundary>);
+        const element = getByText(/Hi Bibo!/i);
+        expect(element).toBeInTheDocument();
     });
 
     it('renders the correct text when a child throws', () => {
-        const c = enzyme.mount(<ErrorBoundary><MyWidget name={null}/></ErrorBoundary>);
-        expect(c.find('.errorBoundary-header').text()).toEqual('Something went wrong.');
-        expect(c.find('.errorBoundary-details').text()).toContain('Oh no!');
+        const {getByText} = render(<ErrorBoundary><MyWidget name={null}/></ErrorBoundary>);
+        const element1 = getByText(/Something went wrong/i);
+        expect(element1).toBeInTheDocument();
+        const element2 = getByText(/Oh no/i);
+        expect(element2).toBeInTheDocument();
     });
 
     it('throws when no children given', () => {
         expect(() => {
-            enzyme.shallow(<ErrorBoundary/>);
+            render(<ErrorBoundary/>);
         }).toThrow();
     });
 });
