@@ -12,8 +12,10 @@ import Box from '@material-ui/core/Box';
 import FormControl from '@material-ui/core/FormControl';
 
 import { Time, TimeRange } from '../model/timeSeries';
+import { TimeAnimationInterval } from '../states/controlState';
 import { WithLocale } from '../util/lang';
 import { I18N } from '../config';
+import { useEffect, useRef } from 'react';
 
 
 // noinspection JSUnusedLocalSymbols
@@ -32,137 +34,133 @@ interface TimePlayerProps extends WithStyles<typeof styles>, WithLocale {
     incSelectedTime: (increment: -1 | 1) => void;
     selectedTimeRange: TimeRange | null;
     timeAnimationActive: boolean;
-    timeAnimationInterval: number;
-    updateTimeAnimation: (active: boolean, interval: number) => void;
+    timeAnimationInterval: TimeAnimationInterval;
+    updateTimeAnimation: (active: boolean, interval: TimeAnimationInterval) => void;
 }
 
 
-class TimePlayer extends React.Component<TimePlayerProps> {
+const TimePlayer: React.FC<TimePlayerProps> = ({
+                                                   classes,
+                                                   timeAnimationActive,
+                                                   timeAnimationInterval,
+                                                   updateTimeAnimation,
+                                                   selectedTime,
+                                                   selectedTimeRange,
+                                                   selectTime,
+                                                   incSelectedTime
+                                               }) => {
 
-    private intervalId: number | null = null;
+    const intervalId = useRef<number | null>(null);
 
-    private handlePlayEvent = () => {
-        this.props.incSelectedTime(1);
+    useEffect(() => {
+        playOrNot();
+        return uninstallTimer;
+    });
+
+    const handlePlayEvent = () => {
+        incSelectedTime(1);
     };
 
-    private handlePlayButtonClick = () => {
-        const {timeAnimationActive, timeAnimationInterval, updateTimeAnimation} = this.props;
+    const handlePlayButtonClick = () => {
         updateTimeAnimation(!timeAnimationActive, timeAnimationInterval);
     };
 
-    private handleNextTimeStepButtonClick = () => {
-        this.props.incSelectedTime(1);
+    const handleNextTimeStepButtonClick = () => {
+        incSelectedTime(1);
     };
 
-    private handlePrevTimeStepButtonClick = () => {
-        this.props.incSelectedTime(-1);
+    const handlePrevTimeStepButtonClick = () => {
+        incSelectedTime(-1);
     };
 
-    private handleFirstTimeStepButtonClick = () => {
-        const {selectedTimeRange} = this.props;
-        this.props.selectTime(selectedTimeRange ? selectedTimeRange[0] : null);
+    const handleFirstTimeStepButtonClick = () => {
+        selectTime(selectedTimeRange ? selectedTimeRange[0] : null);
     };
 
-    private handleLastTimeStepButtonClick = () => {
-        const {selectedTimeRange} = this.props;
-        this.props.selectTime(selectedTimeRange ? selectedTimeRange[1] : null);
+    const handleLastTimeStepButtonClick = () => {
+        selectTime(selectedTimeRange ? selectedTimeRange[1] : null);
     };
 
-    private playOrNot() {
-        if (this.props.timeAnimationActive) {
-            this.installTimer();
+    const playOrNot = () => {
+        if (timeAnimationActive) {
+            installTimer();
         } else {
-            this.uninstallTimer();
+            uninstallTimer();
         }
     };
 
-    private installTimer() {
-        this.uninstallTimer();
-        this.intervalId = window.setInterval(this.handlePlayEvent, this.props.timeAnimationInterval);
+    const installTimer = () => {
+        uninstallTimer();
+        intervalId.current = window.setInterval(handlePlayEvent, timeAnimationInterval);
     };
 
-    private uninstallTimer() {
-        if (this.intervalId !== null) {
-            window.clearInterval(this.intervalId);
-            this.intervalId = null;
+    const uninstallTimer = () => {
+        if (intervalId.current !== null) {
+            window.clearInterval(intervalId.current!);
+            intervalId.current = null;
         }
     };
 
-    componentDidMount(): void {
-        this.playOrNot();
-    }
 
-    componentDidUpdate(): void {
-        this.playOrNot();
-    }
+    const isValid = typeof selectedTime === 'number';
 
-    componentWillUnmount(): void {
-        this.uninstallTimer();
-    }
+    const playToolTip = I18N.get(timeAnimationActive ? 'Stop' : 'Start');
+    const playIcon = timeAnimationActive ? <PauseCircleOutline/> : <PlayCircleOutline/>;
 
-    render() {
-        let {classes, selectedTime, timeAnimationActive} = this.props;
+    const playButton = (
+        <IconButton
+            disabled={!isValid}
+            aria-label={playToolTip}
+            onClick={handlePlayButtonClick}
+        >
+            {playIcon}
+        </IconButton>
+    );
 
-        const isValid = typeof selectedTime === 'number';
+    const firstTimeStepButton = (
+        <IconButton
+            disabled={timeAnimationActive}
+            onClick={handleFirstTimeStepButtonClick}
+        >
+            <FirstPageIcon/>
+        </IconButton>
+    );
 
-        const playToolTip = I18N.get(timeAnimationActive ? 'Stop' : 'Start');
-        const playIcon = timeAnimationActive ? <PauseCircleOutline/> : <PlayCircleOutline/>;
+    const prevTimeStepButton = (
+        <IconButton
+            disabled={timeAnimationActive}
+            onClick={handlePrevTimeStepButtonClick}
+        >
+            <ChevronLeftIcon/>
+        </IconButton>
+    );
+    const nextTimeStepButton = (
+        <IconButton
+            disabled={timeAnimationActive}
+            onClick={handleNextTimeStepButtonClick}
+        >
+            <ChevronRightIcon/>
+        </IconButton>
+    );
+    const lastTimeStepButton = (
+        <IconButton
+            disabled={timeAnimationActive}
+            onClick={handleLastTimeStepButtonClick}
+        >
+            <LastPageIcon/>
+        </IconButton>);
 
-        const playButton = (
-            <IconButton
-                disabled={!isValid}
-                aria-label={playToolTip}
-                onClick={this.handlePlayButtonClick}
-            >
-                {playIcon}
-            </IconButton>
-        );
-
-        const firstTimeStepButton = (
-            <IconButton
-                disabled={timeAnimationActive}
-                onClick={this.handleFirstTimeStepButtonClick}
-            >
-                <FirstPageIcon/>
-            </IconButton>
-        );
-
-        const prevTimeStepButton = (
-            <IconButton
-                disabled={timeAnimationActive}
-                onClick={this.handlePrevTimeStepButtonClick}
-            >
-                <ChevronLeftIcon/>
-            </IconButton>
-        );
-        const nextTimeStepButton = (
-            <IconButton
-                disabled={timeAnimationActive}
-                onClick={this.handleNextTimeStepButtonClick}
-            >
-                <ChevronRightIcon/>
-            </IconButton>
-        );
-        const lastTimeStepButton = (
-            <IconButton
-                disabled={timeAnimationActive}
-                onClick={this.handleLastTimeStepButtonClick}
-            >
-                <LastPageIcon/>
-            </IconButton>);
-
-        return (
-            <FormControl className={classes.formControl}>
-                <Box>
-                    {firstTimeStepButton}
-                    {prevTimeStepButton}
-                    {playButton}
-                    {nextTimeStepButton}
-                    {lastTimeStepButton}
-                </Box>
-            </FormControl>
-        );
-    }
-}
+    return (
+        <FormControl className={classes.formControl}>
+            <Box>
+                {firstTimeStepButton}
+                {prevTimeStepButton}
+                {playButton}
+                {nextTimeStepButton}
+                {lastTimeStepButton}
+            </Box>
+        </FormControl>
+    );
+};
 
 export default withStyles(styles)(TimePlayer);
