@@ -5,7 +5,7 @@ import { Theme } from '@material-ui/core';
 import { WithLocale } from '../util/lang';
 import TimeSeriesChart from './TimeSeriesChart';
 import { Time, TimeRange, TimeSeriesGroup } from '../model/timeSeries';
-import { Place, PlaceInfo } from "../model/place";
+import { Place, PlaceInfo } from '../model/place';
 import TimeRangeSlider from './TimeRangeSlider';
 
 
@@ -47,59 +47,56 @@ interface TimeSeriesChartsProps extends WithStyles<typeof styles>, WithLocale {
     places: Place[];
 }
 
-class TimeSeriesCharts extends React.Component<TimeSeriesChartsProps> {
+const TimeSeriesCharts: React.FC<TimeSeriesChartsProps> = ({
+                                                               classes, locale,
+                                                               timeSeriesGroups,
+                                                               selectedTime, selectedTimeRange,
+                                                               dataTimeRange, selectTime, selectTimeRange,
+                                                               removeTimeSeriesGroup, showPointsOnly, showErrorBars,
+                                                               placeInfos, places, selectPlace
+                                                           }) => {
 
-    render() {
-        const {
-            classes, locale,
-            timeSeriesGroups,
-            selectedTime, selectedTimeRange,
-            dataTimeRange, selectTime, selectTimeRange,
-            removeTimeSeriesGroup, showPointsOnly, showErrorBars,
-            placeInfos, places, selectPlace
-        } = this.props;
 
-        const charts = timeSeriesGroups.map((timeSeriesGroup: TimeSeriesGroup) => {
-            const completed = timeSeriesGroup.timeSeriesArray.map(item => (
-                item.dataProgress ? 100 * item.dataProgress : 0
-            ));
-
-            return (
-                <TimeSeriesChart
-                    key={timeSeriesGroup.id}
-                    locale={locale}
-                    timeSeriesGroup={timeSeriesGroup}
-                    selectedTime={selectedTime}
-                    selectedTimeRange={selectedTimeRange}
-                    dataTimeRange={dataTimeRange}
-                    selectTime={selectTime}
-                    selectTimeRange={selectTimeRange}
-                    removeTimeSeriesGroup={removeTimeSeriesGroup}
-                    completed={completed}
-                    showPointsOnly={showPointsOnly}
-                    showErrorBars={showErrorBars}
-                    placeInfos={placeInfos}
-                    places={places}
-                    selectPlace={selectPlace}
-                />
-            );
-        });
-
-        if (charts.length === 0) {
-            return null;
-        }
+    const charts = timeSeriesGroups.map((timeSeriesGroup: TimeSeriesGroup) => {
+        const completed = timeSeriesGroup.timeSeriesArray.map(item => (
+            item.dataProgress ? 100 * item.dataProgress : 0
+        ));
 
         return (
-            <div className={classes.chartContainer}>
-                <TimeRangeSlider
-                    selectedTimeRange={selectedTimeRange}
-                    dataTimeRange={dataTimeRange}
-                    selectTimeRange={selectTimeRange}
-                />
-                {charts}
-            </div>
+            <TimeSeriesChart
+                key={timeSeriesGroup.id}
+                locale={locale}
+                timeSeriesGroup={timeSeriesGroup}
+                selectedTime={selectedTime}
+                selectedTimeRange={selectedTimeRange}
+                dataTimeRange={dataTimeRange}
+                selectTime={selectTime}
+                selectTimeRange={selectTimeRange}
+                removeTimeSeriesGroup={removeTimeSeriesGroup}
+                completed={completed}
+                showPointsOnly={showPointsOnly}
+                showErrorBars={showErrorBars}
+                placeInfos={placeInfos}
+                places={places}
+                selectPlace={selectPlace}
+            />
         );
+    });
+
+    if (charts.length === 0) {
+        return null;
     }
-}
+
+    return (
+        <div className={classes.chartContainer}>
+            <TimeRangeSlider
+                selectedTimeRange={selectedTimeRange}
+                dataTimeRange={dataTimeRange}
+                selectTimeRange={selectTimeRange}
+            />
+            {charts}
+        </div>
+    );
+};
 
 export default withStyles(styles, {withTheme: true})(TimeSeriesCharts);
