@@ -5,13 +5,9 @@ import { Theme, WithStyles, createStyles, withStyles } from '@material-ui/core';
 import SettingsApplicationsIcon from '@material-ui/icons/SettingsApplications';
 import deepOrange from '@material-ui/core/colors/deepOrange';
 import { AppBar, Toolbar, Typography, IconButton } from '@material-ui/core';
-// import { AppBar, Toolbar, Typography, IconButton, Badge } from '@material-ui/core';
-// import NotificationsIcon from '@material-ui/icons/Notifications';
-// import Avatar from '@material-ui/core/Avatar';
 
 import { AppState } from '../states/appState';
-import logo from '../resources/logo.png';
-import { VIEWER_LOGO_WIDTH, VIEWER_HEADER_BACKGROUND_COLOR, VIEWER_APP_NAME } from '../config';
+import { getBranding } from '../config';
 import ServerDialog from './ServerDialog';
 import SettingsDialog from './SettingsDialog';
 import { openDialog } from '../actions/controlActions';
@@ -27,7 +23,7 @@ interface AppBarProps extends WithStyles<typeof styles> {
 const mapStateToProps = (state: AppState) => {
     return {
         locale: state.controlState.locale,
-        appName: VIEWER_APP_NAME,
+        appName: getBranding().appBarTitle,
     };
 };
 
@@ -39,7 +35,8 @@ const mapDispatchToProps = {
 const styles = (theme: Theme) => createStyles(
     {
         toolbar: {
-            backgroundColor: VIEWER_HEADER_BACKGROUND_COLOR,
+            backgroundColor: getBranding().headerBackgroundColor,
+            paddingRight: theme.spacing(1),
         },
         toolbarIcon: {
             display: 'flex',
@@ -69,39 +66,35 @@ const styles = (theme: Theme) => createStyles(
         },
     });
 
-class _AppBar extends React.Component<AppBarProps> {
+const _AppBar: React.FC<AppBarProps> = ({classes, appName, openDialog}) => {
 
-    handleSettingsButtonClicked = (event: React.MouseEvent<HTMLButtonElement>) => {
-        this.props.openDialog('settings');
+    const handleSettingsButtonClicked = () => {
+        openDialog('settings');
     };
 
-    render() {
-        const {classes, appName} = this.props;
-        return (
-            <AppBar
-                position="absolute"
-                className={classNames(classes.appBar)}
-            >
-                <Toolbar disableGutters={true} className={classes.toolbar}>
-                    <img src={logo} alt={'xcube logo'} width={VIEWER_LOGO_WIDTH} className={classes.logo}/>
-                    <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
-                        {appName}
-                    </Typography>
-                    {/*<IconButton>*/}
-                    {/*<Badge badgeContent={4} color={"secondary"}>*/}
-                    {/*<NotificationsIcon/>*/}
-                    {/*</Badge>*/}
-                    {/*</IconButton>*/}
-                    {/*<Avatar className={classes.orangeAvatar}>CL</Avatar>*/}
-                    <IconButton onClick={this.handleSettingsButtonClicked}>
-                        <SettingsApplicationsIcon/>
-                    </IconButton>
-                </Toolbar>
-                <ServerDialog/>
-                <SettingsDialog/>
-            </AppBar>
-        );
-    }
-}
+    return (
+        <AppBar
+            position="absolute"
+            className={classNames(classes.appBar)}
+        >
+            <Toolbar disableGutters={true} className={classes.toolbar}>
+                <img
+                    src={getBranding().logoPath}
+                    width={getBranding().logoWidth}
+                    alt={'xcube logo'}
+                    className={classes.logo}
+                />
+                <Typography component="h1" variant="h6" color="inherit" noWrap className={classes.title}>
+                    {appName}
+                </Typography>
+                <IconButton onClick={handleSettingsButtonClicked}>
+                    <SettingsApplicationsIcon/>
+                </IconButton>
+            </Toolbar>
+            <ServerDialog/>
+            <SettingsDialog/>
+        </AppBar>
+    );
+};
 
 export default connect(mapStateToProps, mapDispatchToProps)(withStyles(styles)(_AppBar));
