@@ -4,7 +4,7 @@ import { AppState } from '../states/appState';
 import SettingsDialog from '../components/SettingsDialog';
 import { changeLocale, closeDialog, openDialog, updateSettings } from '../actions/controlActions';
 import { selectedServerSelector } from "../selectors/controlSelectors";
-import { VIEWER_VERSION } from '../config';
+import { getVersion } from '../config';
 
 
 const mapStateToProps = (state: AppState) => {
@@ -13,7 +13,7 @@ const mapStateToProps = (state: AppState) => {
         open: state.controlState.dialogOpen['settings'],
         settings: state.controlState,
         selectedServer: selectedServerSelector(state),
-        viewerVersion: VIEWER_VERSION,
+        viewerVersion: getVersion(),
         serverInfo: state.dataState.serverInfo,
     };
 };
