@@ -3,8 +3,8 @@ import { MessageLogState, newMessageLogState } from '../states/messageLogState';
 import { HIDE_MESSAGE, MessageLogAction, POST_MESSAGE } from "../actions/messageLogActions";
 
 
-export function messageLogReducer(state: MessageLogState, action: MessageLogAction): MessageLogState  {
-    if (typeof state === 'undefined') {
+export function messageLogReducer(state: MessageLogState | undefined, action: MessageLogAction): MessageLogState  {
+    if (state === undefined) {
         state = newMessageLogState();
     }
     switch (action.type) {
@@ -17,7 +17,7 @@ export function messageLogReducer(state: MessageLogState, action: MessageLogActi
             return {...state, newEntries: [messageLogEntry, ...state.newEntries]}
         }
         case HIDE_MESSAGE: {
-            const i = state.newEntries.findIndex(e => e.id == action.messageId);
+            const i = state.newEntries.findIndex(e => e.id === action.messageId);
             if (i >= 0) {
                 const hiddenEntry = state.newEntries[i];
                 const newEntries = [...state.newEntries];
