@@ -1,8 +1,11 @@
-import {  OlGeometry , OlExtent, OlBaseObject } from '../components/ol/types';
+import { default as OlBaseObject } from 'ol/Object';
+import { Geometry as OlGeometry } from 'ol/geom';
+import { Extent as OlExtent } from 'ol/extent';
 
 import { Time, TimeRange } from '../model/timeSeries';
-import { VIEWER_DEFAULT_API_SERVER } from '../config';
 import { loadUserSettings } from './userSettings';
+import { getBranding, getDefaultApiServer } from '../config';
+
 
 export type TimeAnimationInterval = 250 | 500 | 1000 | 2500;
 export const TIME_ANIMATION_INTERVALS: TimeAnimationInterval[] = [250, 500, 1000, 2500];
@@ -31,17 +34,19 @@ export interface ControlState {
     legalAgreementAccepted: boolean;
     mapInteraction: MapInteraction;
     imageSmoothingEnabled: boolean;
+    baseMapUrl: string;
 }
 
 
 export function newControlState(): ControlState {
+    const branding = getBranding();
     const state: ControlState = {
         selectedDatasetId: 'local',
         selectedVariableName: 'conc_chl',
         selectedPlaceGroupIds: ['user'],
         selectedPlaceId: null,
         selectedUserPlaceId: null,
-        selectedServerId: VIEWER_DEFAULT_API_SERVER.id,
+        selectedServerId: getDefaultApiServer().id,
         selectedTime: null,
         selectedTimeRange: null,
         timeSeriesUpdateMode: 'add',
@@ -57,6 +62,7 @@ export function newControlState(): ControlState {
         legalAgreementAccepted: false,
         mapInteraction: 'Point',
         imageSmoothingEnabled: false,
+        baseMapUrl: branding.baseMapUrl || 'http://a.tile.osm.org/{z}/{x}/{y}.png',
     };
     return loadUserSettings(state);
 }
