@@ -1,3 +1,4 @@
+import { getBrandingName } from '../config';
 import { Storage, getLocalStorage } from '../util/storage';
 import { Server } from '../model/server';
 import { ControlState } from './controlState';
@@ -41,6 +42,7 @@ export function storeUserSettings(settings: ControlState) {
             _storeProperty(storage, 'showTimeSeriesPointsOnly', settings);
             _storeProperty(storage, 'timeAnimationInterval', settings);
             _storeProperty(storage, 'imageSmoothingEnabled', settings);
+            _storeProperty(storage, 'baseMapUrl', settings);
         } catch (e) {
             console.warn(`failed to store user settings: ${e}`);
         }
@@ -58,6 +60,7 @@ export function loadUserSettings(defaultSettings: ControlState): ControlState {
             _loadBooleanProperty(storage, 'showTimeSeriesPointsOnly', settings, defaultSettings);
             _loadIntProperty(storage, 'timeAnimationInterval', settings, defaultSettings);
             _loadBooleanProperty(storage, 'imageSmoothingEnabled', settings, defaultSettings);
+            _loadStringProperty(storage, 'baseMapUrl', settings, defaultSettings);
         } catch (e) {
             console.warn(`failed to load user settings: ${e}`);
         }
@@ -68,27 +71,34 @@ export function loadUserSettings(defaultSettings: ControlState): ControlState {
 
 
 function _storeProperty(storage: Storage, propertyName: string, source: any) {
-    storage.setItem(`xcube.${propertyName}`, source[propertyName] + '');
+    const brandingName = getBrandingName();
+    storage.setItem(`xcube.${brandingName}.${propertyName}`, source[propertyName] + '');
     // console.log(`stored xcube.${propertyName}`, source);
 }
 
 function _loadBooleanProperty(storage: Storage, propertyName: string, target: any, defaultObj: any) {
-    const value = storage.getItem(`xcube.${propertyName}`);
-    if (value !== null) {
-        target[propertyName] = value == 'true';
-    } else {
-        target[propertyName] = !!defaultObj[propertyName];
-    }
-    // console.log(`loaded xcube.${propertyName}`, target);
+    _loadProperty(storage, propertyName, target, defaultObj, (value) => value === 'true');
 }
 
 function _loadIntProperty(storage: Storage, propertyName: string, target: any, defaultObj: any) {
-    const value = storage.getItem(`xcube.${propertyName}`);
+    _loadProperty(storage, propertyName, target, defaultObj, parseInt);
+}
+
+function _loadStringProperty(storage: Storage, propertyName: string, target: any, defaultObj: any) {
+    _loadProperty(storage, propertyName, target, defaultObj, (value) => value);
+}
+
+function _loadProperty(storage: Storage,
+                       propertyName: string,
+                       target: any,
+                       defaultObj: any,
+                       converter: (value: string) => any) {
+    const brandingName = getBrandingName();
+    const value = storage.getItem(`xcube.${brandingName}.${propertyName}`);
     if (value !== null) {
-        target[propertyName] = parseInt(value);
+        target[propertyName] = converter(value);
     } else {
         target[propertyName] = defaultObj[propertyName];
     }
     // console.log(`loaded xcube.${propertyName}`, target);
 }
-
